+++ src/lib/Primitive.ts
@@ -1,4 +1,0 @@
-const enum Primitive {
-	Int, Float, Boolean, String
-}
-export default Primitive